Adding Auto-Save functionality
Prototype for auto-save works I have it set to cap out the max number of files it'll keep on auto-save (since they have a timestamp attached to the name it won't just overwrite).

My idea was if we trigger the auto-save every 3 minutes then if someone really gets into the groove of editing or adding to their library manually and loses track of time, cap 20 files * 3 min will give them backups for their whole last hour of work at any given time. They'll only ever lose their last 3 minutes of work in power loss.

It'll auto-delete the oldest ones when it reaches the cap and auto-saves again

I tested it by changing the trigger time to 10 sec and then uploading something new every 10 sec / adding a manual entry.

Currently there is no user interaction necessary as it creates a folder in the system agnostic default path for a users home folder called "GameLoom Exports" and puts them in there (I added Exports after I took the screenshot below)

The timer for it runs as a background daemon so it gets put onto a daemon thread for efficiency
https://docs.oracle.com/javase/8/docs/api/java/util/Timer.html

The hash is a modest MD5. Since we aren't actually using it for encryption we can go with a simple hashing cipher that won't use much resources but still maintain accuracy. MD5 is a common internet standard you'll see listed on downloads for security software or operating systems a lot. The hash is based on the contents of the library ArrayList object not the VBox object (but they're one in the same anyway)