Added Popular App Filtering
Updated Normalizer and GameCSVImporter to filter out the top 50 most popular apps for PS4/PS5 that aren't games.

Had to add an if statement into the importer to rule them out but put the array of app names in Normalizer.
Probably best that all those long lists like that stay in that one file