Updated Scroll Window and toString()
Updated the scrollable window to scroll up,down.left,right.

Changed the toString() method to not be locked into the pre-defined 5 fields instead it has those 5 preferred and will list them first if the game has those keys but if it has an N/A value or equivalent it wont display it on toString() (in the visible window, all the fields are still there in the backend). After it shows the preferred ones it shows whatever else the game has for attributes and it capitalizes the first letter of each normalized key and replaces _ with spaces for displaying (doesn't actually change the normalized values just displays them this way)