35 scrum 31 help tab finalization (#46)
* Contact Support Link Functional

Made Contact Support Link Functional on the Help Tab

* Updating zip file contents + README

Swapping the version of the zip included in the repo, the public facing one doesn't need our test files and it needed the updated README. I have put a placeholder FAQ in there until I finish it and swap it back out again. I have gone ahead and made the txt version of the readme and it is included in the zip. I am addressing the other user instruction formatting with a video tutorial.