35 help tab finalization (#68)
* Added Working Link to FAQ

Made the link direct to the 'main' branch rather than the development branch. I figured we'll use 'main' for versioned releases.

* Added User's Guide Playlist Link

Added the working YouTube link to the Playlist that serves as the User's Guide. We can consider making a written version as well for the WIKI but I know most people are more partial to videos so I figured we start there.

* Repackaged Jar & Zip

Repackaged JAR and ZIP for YouTube Recording. Going to knock it all out today.