Fixing Manual Entry Voiding Invalids
Previously, if you submit 3 entries (any # really) but only 1 of them is valid (or any other number besides the total) the manual entry tab would add the valid ones and basically void the invalid ones that didn't meet the 3 required fields.

I've changed it to where it'll leave the invalid ones behind and only submit the valid ones and then just highlight the missing value boxes in a red outline instead of just sending them to void and making the user have to click out of the tab to check which ones didn't take.