Update Title Quote Consistency in Edit Tab
Noticed somewhere along the line I lost some code that handled appending quotes around all of the manual entry and edit tab stuff that the user can enter so I've fixed that.

I think what happened was at one point the merge conflicts were too much and I decided to just make a new branch off of development and redo the changes I had made up to that point (around the FAQ stuff) and I must've also had that in that branch and forgot to carry it over. It's added back in now